Move account-related commands to eth_accounts.go
